graphQL을 적용해서 문제 제출횟수 구하는 부분 구현(정상동작 확인)

+++ problem-service/src/main/kotlin/com/sevity/problemservice/domain/Submission.kt
@@ -5,10 +5,10 @@ import javax.persistence.*
 @Table(name = "submissions")
 data class Submission(
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
+    val id: Int = 0,
 
     @Column(name = "problem_id", nullable = false)
-    val problemId: Long,
+    val problemId: Int,
 
     @Column(name = "user_id", nullable = false)
     val userId: Long,
